pyproject.toml
[tool.poetry]
name = "boardtt"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^11.0.0"
pytesseract = "^0.3.13"

[tool.poetry.group.dev.dependencies]
ruff = "^0.7.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
